Remove public modifier from RequestConfigHolder. Move WithRequestCallable in the same package with RequestConfigHolder.
