+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/RequestConfigHolder.java
@@ -1,18 +1,18 @@
 package io.fabric8.kubernetes.client;
 
-public class RequestConfigHolder {
+class RequestConfigHolder {
 
   private static final ThreadLocal<RequestConfig> REQUEST_CONFIG = new ThreadLocal<>();
 
-  public static RequestConfig get() {
+  static RequestConfig get() {
     return REQUEST_CONFIG.get();
   }
 
-  public static void set(RequestConfig requestConfig) {
+  static void set(RequestConfig requestConfig) {
     REQUEST_CONFIG.set(requestConfig);
   }
 
-  public static void remove() {
+  static void remove() {
     REQUEST_CONFIG.remove();
   }
 }
